requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bing-image-urls==0.1.5
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
exceptiongroup==1.1.3
fonttools==4.42.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.276
langsmith==0.0.26
logzero==1.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.9
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
plotly==5.16.1
protobuf==4.24.1
pyarrow==13.0.0
pydantic==2.2.1
pydantic_core==2.6.1
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
rich==13.5.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
streamlit==1.25.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
yarl==1.9.2
zipp==3.16.2
python-dotenv==1.0.0
